pyproject.toml
[tool.poetry]
name = "python-postgresql"
version = "0.1.0"
description = ""
authors = ["manuev2 <manny.4163@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "^2.9.9"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
